Add first few Great House pages.
Each Great House gets a little blurb on its own page to show off what
that House is about. I am debating whether or not to exclude them from
the nav display; they might be better off as little links. For now, I'll
add them all and then, _after_ that, make the decision as to whether or
not to exclude them.