Recolor house mons to white
The original black House Mons were too dark to show up clearly on the
dark color theme. The dark color theme goes well with LCA so I don't
want to change that. The remaining mons are white, which will contrast
nicely with the dark background and hopefully make them visible.