Add remaining House pages
Each one has the same template and the same image, so they're easy. I'm
still not sure whether or not I want them to show up in the nav display;
that's 10 items that might not actually _need_ to be in the sidebar.

Let's see how this looks before making that decision.