Adjust nav_exclude for House pages
Excluding them from the nav bar helps keep the nav bar clean, and
adjusting the nav order makes sure they display in the right order.

Adjusting the nav order is not strictly required because the nav order
will default to alphabetical (the same order as we want), but explicit
is better than implicit.