Add content for Gear page
The explanation for the Resources system is the best I can do. Although
I like the narrative way that Exalted handles money, it can be kind of
hard to explain to people coming from, say, D&D. I hope that the five
paragraphs are enough to convey how it's supposed to work.